chore: exclude .ruby-lsp folder from RuboCop checks
The .ruby-lsp folder is auto-generated by the Ruby LSP and should
not be checked by RuboCop as it's not part of the project code.
