+++ test/openai_integration_test.rb
@@ -730,194 +730,10 @@ module OpenAI
       mock_responses_api.verify
       mock_mcp_client.verify
     end
-  end
 
-  # Temperature parameter tests
+    # Temperature parameter tests
 
-  def test_chat_completion_with_temperature_for_gpt_model
-    # Create a logger that accepts anything
-    mock_logger = Minitest::Mock.new
-    def mock_logger.info(*args); end
-    def mock_logger.error(*args); end
-    def mock_logger.log(*args); end
-
-    mock_mcp_client = Minitest::Mock.new
-    mock_openai_client = Minitest::Mock.new
-
-    api = ClaudeSwarm::OpenAIChatCompletion.new(
-      openai_client: mock_openai_client,
-      mcp_client: mock_mcp_client,
-      available_tools: [],
-      logger: mock_logger,
-      instance_name: "test",
-      model: "gpt-4",
-      temperature: 0.8,
-    )
-
-    # Mock the chat call - verify temperature is included for GPT models
-    mock_response = {
-      "choices" => [{
-        "message" => {
-          "role" => "assistant",
-          "content" => "Hello with temperature!",
-        },
-      }],
-    }
-
-    mock_openai_client.expect(:chat, mock_response) do |params|
-      params[:parameters][:temperature] == 0.8 &&
-        params[:parameters][:model] == "gpt-4"
-    end
-
-    result = api.execute("Hello")
-
-    assert_equal("Hello with temperature!", result)
-
-    mock_openai_client.verify
-  end
-
-  def test_chat_completion_without_temperature_for_o_series_model
-    # Create a logger that accepts anything
-    mock_logger = Minitest::Mock.new
-    def mock_logger.info(*args); end
-    def mock_logger.error(*args); end
-    def mock_logger.log(*args); end
-
-    mock_mcp_client = Minitest::Mock.new
-    mock_openai_client = Minitest::Mock.new
-
-    api = ClaudeSwarm::OpenAIChatCompletion.new(
-      openai_client: mock_openai_client,
-      mcp_client: mock_mcp_client,
-      available_tools: [],
-      logger: mock_logger,
-      instance_name: "test",
-      model: "o1",
-      temperature: 0.8, # This should be ignored
-    )
-
-    # Mock the chat call - verify temperature is NOT included for O series models
-    mock_response = {
-      "choices" => [{
-        "message" => {
-          "role" => "assistant",
-          "content" => "Hello without temperature!",
-        },
-      }],
-    }
-
-    mock_openai_client.expect(:chat, mock_response) do |params|
-      !params[:parameters].key?(:temperature) &&
-        params[:parameters][:model] == "o1"
-    end
-
-    result = api.execute("Hello")
-
-    assert_equal("Hello without temperature!", result)
-
-    mock_openai_client.verify
-  end
-
-  def test_responses_api_with_temperature_for_gpt_model
-    # Create a logger that accepts anything
-    mock_logger = Minitest::Mock.new
-    def mock_logger.info(*args); end
-    def mock_logger.error(*args); end
-    def mock_logger.log(*args); end
-
-    mock_mcp_client = Minitest::Mock.new
-    mock_openai_client = Minitest::Mock.new
-    mock_responses_api = Minitest::Mock.new
-
-    # Set up OpenAI client to return responses API
-    mock_openai_client.expect(:responses, mock_responses_api)
-
-    api = ClaudeSwarm::OpenAIResponses.new(
-      openai_client: mock_openai_client,
-      mcp_client: mock_mcp_client,
-      available_tools: [],
-      logger: mock_logger,
-      instance_name: "test",
-      model: "gpt-4o",
-      temperature: 0.7,
-    )
-
-    # Mock the create call - verify temperature is included for GPT models
-    mock_response = {
-      "id" => "resp_123",
-      "output" => [{
-        "type" => "message",
-        "content" => [{
-          "type" => "text",
-          "text" => "Hello from responses API with temperature!",
-        }],
-      }],
-    }
-
-    mock_responses_api.expect(:create, mock_response) do |params|
-      params[:parameters][:temperature] == 0.7 &&
-        params[:parameters][:model] == "gpt-4o"
-    end
-
-    result = api.execute("Hello")
-
-    assert_equal("Hello from responses API with temperature!", result)
-
-    mock_openai_client.verify
-    mock_responses_api.verify
-  end
-
-  def test_responses_api_without_temperature_for_o_series_model
-    # Create a logger that accepts anything
-    mock_logger = Minitest::Mock.new
-    def mock_logger.info(*args); end
-    def mock_logger.error(*args); end
-    def mock_logger.log(*args); end
-
-    mock_mcp_client = Minitest::Mock.new
-    mock_openai_client = Minitest::Mock.new
-    mock_responses_api = Minitest::Mock.new
-
-    # Set up OpenAI client to return responses API
-    mock_openai_client.expect(:responses, mock_responses_api)
-
-    api = ClaudeSwarm::OpenAIResponses.new(
-      openai_client: mock_openai_client,
-      mcp_client: mock_mcp_client,
-      available_tools: [],
-      logger: mock_logger,
-      instance_name: "test",
-      model: "o3",
-      temperature: 0.9, # This should be ignored
-    )
-
-    # Mock the create call - verify temperature is NOT included for O series models
-    mock_response = {
-      "id" => "resp_123",
-      "output" => [{
-        "type" => "message",
-        "content" => [{
-          "type" => "text",
-          "text" => "Hello from responses API without temperature!",
-        }],
-      }],
-    }
-
-    mock_responses_api.expect(:create, mock_response) do |params|
-      !params[:parameters].key?(:temperature) &&
-        params[:parameters][:model] == "o3"
-    end
-
-    result = api.execute("Hello")
-
-    assert_equal("Hello from responses API without temperature!", result)
-
-    mock_openai_client.verify
-    mock_responses_api.verify
-  end
-
-  def test_chat_completion_all_o_series_models_exclude_temperature
-    ["o1", "o1 Preview", "o1-mini", "o1-pro", "o3", "o3-mini", "o3-pro", "o3-deep-research", "o4-mini", "o4-mini-deep-research"].each do |model|
+    def test_chat_completion_with_temperature_for_gpt_model
       # Create a logger that accepts anything
       mock_logger = Minitest::Mock.new
       def mock_logger.info(*args); end
@@ -927,41 +743,81 @@ module OpenAI
       mock_mcp_client = Minitest::Mock.new
       mock_openai_client = Minitest::Mock.new
 
-      api = ClaudeSwarm::OpenAIChatCompletion.new(
+      api = ClaudeSwarm::OpenAI::ChatCompletion.new(
         openai_client: mock_openai_client,
         mcp_client: mock_mcp_client,
         available_tools: [],
         logger: mock_logger,
         instance_name: "test",
-        model: model,
-        temperature: 0.5,
+        model: "gpt-4",
+        temperature: 0.8,
       )
 
-      # Mock the chat call
+      # Mock the chat call - verify temperature is included for GPT models
       mock_response = {
         "choices" => [{
           "message" => {
             "role" => "assistant",
-            "content" => "Response from #{model}",
+            "content" => "Hello with temperature!",
+          },
+        }],
+      }
+
+      mock_openai_client.expect(:chat, mock_response) do |params|
+        params[:parameters][:temperature] == 0.8 &&
+          params[:parameters][:model] == "gpt-4"
+      end
+
+      result = api.execute("Hello")
+
+      assert_equal("Hello with temperature!", result)
+
+      mock_openai_client.verify
+    end
+
+    def test_chat_completion_without_temperature_for_o_series_model
+      # Create a logger that accepts anything
+      mock_logger = Minitest::Mock.new
+      def mock_logger.info(*args); end
+      def mock_logger.error(*args); end
+      def mock_logger.log(*args); end
+
+      mock_mcp_client = Minitest::Mock.new
+      mock_openai_client = Minitest::Mock.new
+
+      api = ClaudeSwarm::OpenAI::ChatCompletion.new(
+        openai_client: mock_openai_client,
+        mcp_client: mock_mcp_client,
+        available_tools: [],
+        logger: mock_logger,
+        instance_name: "test",
+        model: "o1",
+        temperature: 0.8, # This should be ignored
+      )
+
+      # Mock the chat call - verify temperature is NOT included for O series models
+      mock_response = {
+        "choices" => [{
+          "message" => {
+            "role" => "assistant",
+            "content" => "Hello without temperature!",
           },
         }],
       }
 
       mock_openai_client.expect(:chat, mock_response) do |params|
         !params[:parameters].key?(:temperature) &&
-          params[:parameters][:model] == model
+          params[:parameters][:model] == "o1"
       end
 
-      result = api.execute("Test")
+      result = api.execute("Hello")
 
-      assert_equal("Response from #{model}", result)
+      assert_equal("Hello without temperature!", result)
 
       mock_openai_client.verify
     end
-  end
 
-  def test_responses_api_all_o_series_models_exclude_temperature
-    ["o1", "o1 Preview", "o1-mini", "o1-pro", "o3", "o3-mini", "o3-pro", "o3-deep-research", "o4-mini", "o4-mini-deep-research"].each do |model|
+    def test_responses_api_with_temperature_for_gpt_model
       # Create a logger that accepts anything
       mock_logger = Minitest::Mock.new
       def mock_logger.info(*args); end
@@ -975,222 +831,366 @@ module OpenAI
       # Set up OpenAI client to return responses API
       mock_openai_client.expect(:responses, mock_responses_api)
 
-      api = ClaudeSwarm::OpenAIResponses.new(
+      api = ClaudeSwarm::OpenAI::Responses.new(
         openai_client: mock_openai_client,
         mcp_client: mock_mcp_client,
         available_tools: [],
         logger: mock_logger,
         instance_name: "test",
-        model: model,
-        temperature: 0.5,
+        model: "gpt-4o",
+        temperature: 0.7,
       )
 
-      # Mock the create call
+      # Mock the create call - verify temperature is included for GPT models
       mock_response = {
-        "id" => "resp_#{model}",
+        "id" => "resp_123",
         "output" => [{
           "type" => "message",
           "content" => [{
             "type" => "text",
-            "text" => "Response from #{model}",
+            "text" => "Hello from responses API with temperature!",
+          }],
+        }],
+      }
+
+      mock_responses_api.expect(:create, mock_response) do |params|
+        params[:parameters][:temperature] == 0.7 &&
+          params[:parameters][:model] == "gpt-4o"
+      end
+
+      result = api.execute("Hello")
+
+      assert_equal("Hello from responses API with temperature!", result)
+
+      mock_openai_client.verify
+      mock_responses_api.verify
+    end
+
+    def test_responses_api_without_temperature_for_o_series_model
+      # Create a logger that accepts anything
+      mock_logger = Minitest::Mock.new
+      def mock_logger.info(*args); end
+      def mock_logger.error(*args); end
+      def mock_logger.log(*args); end
+
+      mock_mcp_client = Minitest::Mock.new
+      mock_openai_client = Minitest::Mock.new
+      mock_responses_api = Minitest::Mock.new
+
+      # Set up OpenAI client to return responses API
+      mock_openai_client.expect(:responses, mock_responses_api)
+
+      api = ClaudeSwarm::OpenAI::Responses.new(
+        openai_client: mock_openai_client,
+        mcp_client: mock_mcp_client,
+        available_tools: [],
+        logger: mock_logger,
+        instance_name: "test",
+        model: "o3",
+        temperature: 0.9, # This should be ignored
+      )
+
+      # Mock the create call - verify temperature is NOT included for O series models
+      mock_response = {
+        "id" => "resp_123",
+        "output" => [{
+          "type" => "message",
+          "content" => [{
+            "type" => "text",
+            "text" => "Hello from responses API without temperature!",
           }],
         }],
       }
 
       mock_responses_api.expect(:create, mock_response) do |params|
         !params[:parameters].key?(:temperature) &&
-          params[:parameters][:model] == model
+          params[:parameters][:model] == "o3"
       end
 
-      result = api.execute("Test")
+      result = api.execute("Hello")
 
-      assert_equal("Response from #{model}", result)
+      assert_equal("Hello from responses API without temperature!", result)
 
       mock_openai_client.verify
       mock_responses_api.verify
     end
-  end
 
-  # Tests for reasoning_effort restriction to O series models
+    def test_chat_completion_all_o_series_models_exclude_temperature
+      ["o1", "o1 Preview", "o1-mini", "o1-pro", "o3", "o3-mini", "o3-pro", "o3-deep-research", "o4-mini", "o4-mini-deep-research"].each do |model|
+        # Create a logger that accepts anything
+        mock_logger = Minitest::Mock.new
+        def mock_logger.info(*args); end
+        def mock_logger.error(*args); end
+        def mock_logger.log(*args); end
 
-  def test_chat_completion_with_reasoning_effort_for_gpt_model_excludes_it
-    # Create a logger that accepts anything
-    mock_logger = Minitest::Mock.new
-    def mock_logger.info(*args); end
-    def mock_logger.error(*args); end
-    def mock_logger.log(*args); end
+        mock_mcp_client = Minitest::Mock.new
+        mock_openai_client = Minitest::Mock.new
 
-    mock_mcp_client = Minitest::Mock.new
-    mock_openai_client = Minitest::Mock.new
+        api = ClaudeSwarm::OpenAI::ChatCompletion.new(
+          openai_client: mock_openai_client,
+          mcp_client: mock_mcp_client,
+          available_tools: [],
+          logger: mock_logger,
+          instance_name: "test",
+          model: model,
+          temperature: 0.5,
+        )
 
-    api = ClaudeSwarm::OpenAIChatCompletion.new(
-      openai_client: mock_openai_client,
-      mcp_client: mock_mcp_client,
-      available_tools: [],
-      logger: mock_logger,
-      instance_name: "test",
-      model: "gpt-4",
-      reasoning_effort: "high", # This should be ignored for GPT models
-    )
+        # Mock the chat call
+        mock_response = {
+          "choices" => [{
+            "message" => {
+              "role" => "assistant",
+              "content" => "Response from #{model}",
+            },
+          }],
+        }
 
-    # Mock the chat call - verify reasoning_effort is NOT included for GPT models
-    mock_response = {
-      "choices" => [{
-        "message" => {
-          "role" => "assistant",
-          "content" => "Hello from GPT-4!",
-        },
-      }],
-    }
+        mock_openai_client.expect(:chat, mock_response) do |params|
+          !params[:parameters].key?(:temperature) &&
+            params[:parameters][:model] == model
+        end
 
-    mock_openai_client.expect(:chat, mock_response) do |params|
-      !params[:parameters].key?(:reasoning_effort) &&
-        params[:parameters][:model] == "gpt-4"
+        result = api.execute("Test")
+
+        assert_equal("Response from #{model}", result)
+
+        mock_openai_client.verify
+      end
     end
 
-    result = api.execute("Hello")
+    def test_responses_api_all_o_series_models_exclude_temperature
+      ["o1", "o1 Preview", "o1-mini", "o1-pro", "o3", "o3-mini", "o3-pro", "o3-deep-research", "o4-mini", "o4-mini-deep-research"].each do |model|
+        # Create a logger that accepts anything
+        mock_logger = Minitest::Mock.new
+        def mock_logger.info(*args); end
+        def mock_logger.error(*args); end
+        def mock_logger.log(*args); end
 
-    assert_equal("Hello from GPT-4!", result)
+        mock_mcp_client = Minitest::Mock.new
+        mock_openai_client = Minitest::Mock.new
+        mock_responses_api = Minitest::Mock.new
 
-    mock_openai_client.verify
-  end
+        # Set up OpenAI client to return responses API
+        mock_openai_client.expect(:responses, mock_responses_api)
 
-  def test_responses_api_with_reasoning_effort_for_gpt_model_excludes_it
-    # Create a logger that accepts anything
-    mock_logger = Minitest::Mock.new
-    def mock_logger.info(*args); end
-    def mock_logger.error(*args); end
-    def mock_logger.log(*args); end
+        api = ClaudeSwarm::OpenAI::Responses.new(
+          openai_client: mock_openai_client,
+          mcp_client: mock_mcp_client,
+          available_tools: [],
+          logger: mock_logger,
+          instance_name: "test",
+          model: model,
+          temperature: 0.5,
+        )
 
-    mock_mcp_client = Minitest::Mock.new
-    mock_openai_client = Minitest::Mock.new
-    mock_responses_api = Minitest::Mock.new
+        # Mock the create call
+        mock_response = {
+          "id" => "resp_#{model}",
+          "output" => [{
+            "type" => "message",
+            "content" => [{
+              "type" => "text",
+              "text" => "Response from #{model}",
+            }],
+          }],
+        }
 
-    # Set up OpenAI client to return responses API
-    mock_openai_client.expect(:responses, mock_responses_api)
+        mock_responses_api.expect(:create, mock_response) do |params|
+          !params[:parameters].key?(:temperature) &&
+            params[:parameters][:model] == model
+        end
 
-    api = ClaudeSwarm::OpenAIResponses.new(
-      openai_client: mock_openai_client,
-      mcp_client: mock_mcp_client,
-      available_tools: [],
-      logger: mock_logger,
-      instance_name: "test",
-      model: "gpt-4o",
-      reasoning_effort: "medium", # This should be ignored for GPT models
-    )
+        result = api.execute("Test")
 
-    # Mock the create call - verify reasoning is NOT included for GPT models
-    mock_response = {
-      "id" => "resp_123",
-      "output" => [{
-        "type" => "message",
-        "content" => [{
-          "type" => "text",
-          "text" => "Hello from GPT-4o responses API!",
+        assert_equal("Response from #{model}", result)
+
+        mock_openai_client.verify
+        mock_responses_api.verify
+      end
+    end
+
+    # Tests for reasoning_effort restriction to O series models
+
+    def test_chat_completion_with_reasoning_effort_for_gpt_model_excludes_it
+      # Create a logger that accepts anything
+      mock_logger = Minitest::Mock.new
+      def mock_logger.info(*args); end
+      def mock_logger.error(*args); end
+      def mock_logger.log(*args); end
+
+      mock_mcp_client = Minitest::Mock.new
+      mock_openai_client = Minitest::Mock.new
+
+      api = ClaudeSwarm::OpenAI::ChatCompletion.new(
+        openai_client: mock_openai_client,
+        mcp_client: mock_mcp_client,
+        available_tools: [],
+        logger: mock_logger,
+        instance_name: "test",
+        model: "gpt-4",
+        reasoning_effort: "high", # This should be ignored for GPT models
+      )
+
+      # Mock the chat call - verify reasoning_effort is NOT included for GPT models
+      mock_response = {
+        "choices" => [{
+          "message" => {
+            "role" => "assistant",
+            "content" => "Hello from GPT-4!",
+          },
         }],
-      }],
-    }
+      }
 
-    mock_responses_api.expect(:create, mock_response) do |params|
-      !params[:parameters].key?(:reasoning) &&
-        params[:parameters][:model] == "gpt-4o"
+      mock_openai_client.expect(:chat, mock_response) do |params|
+        !params[:parameters].key?(:reasoning_effort) &&
+          params[:parameters][:model] == "gpt-4"
+      end
+
+      result = api.execute("Hello")
+
+      assert_equal("Hello from GPT-4!", result)
+
+      mock_openai_client.verify
     end
 
-    result = api.execute("Hello")
+    def test_responses_api_with_reasoning_effort_for_gpt_model_excludes_it
+      # Create a logger that accepts anything
+      mock_logger = Minitest::Mock.new
+      def mock_logger.info(*args); end
+      def mock_logger.error(*args); end
+      def mock_logger.log(*args); end
 
-    assert_equal("Hello from GPT-4o responses API!", result)
+      mock_mcp_client = Minitest::Mock.new
+      mock_openai_client = Minitest::Mock.new
+      mock_responses_api = Minitest::Mock.new
 
-    mock_openai_client.verify
-    mock_responses_api.verify
-  end
+      # Set up OpenAI client to return responses API
+      mock_openai_client.expect(:responses, mock_responses_api)
 
-  def test_chat_completion_includes_reasoning_effort_only_for_o_series
-    # Test that reasoning_effort is included for O series model
-    mock_logger = Minitest::Mock.new
-    def mock_logger.info(*args); end
-    def mock_logger.error(*args); end
-    def mock_logger.log(*args); end
+      api = ClaudeSwarm::OpenAI::Responses.new(
+        openai_client: mock_openai_client,
+        mcp_client: mock_mcp_client,
+        available_tools: [],
+        logger: mock_logger,
+        instance_name: "test",
+        model: "gpt-4o",
+        reasoning_effort: "medium", # This should be ignored for GPT models
+      )
 
-    mock_mcp_client = Minitest::Mock.new
-    mock_openai_client = Minitest::Mock.new
-
-    api = ClaudeSwarm::OpenAIChatCompletion.new(
-      openai_client: mock_openai_client,
-      mcp_client: mock_mcp_client,
-      available_tools: [],
-      logger: mock_logger,
-      instance_name: "test",
-      model: "o3",
-      reasoning_effort: "low",
-    )
-
-    mock_response = {
-      "choices" => [{
-        "message" => {
-          "role" => "assistant",
-          "content" => "O3 response with reasoning",
-        },
-      }],
-    }
-
-    mock_openai_client.expect(:chat, mock_response) do |params|
-      params[:parameters][:reasoning_effort] == "low" &&
-        params[:parameters][:model] == "o3" &&
-        !params[:parameters].key?(:temperature) # Also verify no temperature
-    end
-
-    result = api.execute("Test")
-
-    assert_equal("O3 response with reasoning", result)
-
-    mock_openai_client.verify
-  end
-
-  def test_responses_api_includes_reasoning_only_for_o_series
-    # Test that reasoning is included for O series model
-    mock_logger = Minitest::Mock.new
-    def mock_logger.info(*args); end
-    def mock_logger.error(*args); end
-    def mock_logger.log(*args); end
-
-    mock_mcp_client = Minitest::Mock.new
-    mock_openai_client = Minitest::Mock.new
-    mock_responses_api = Minitest::Mock.new
-
-    mock_openai_client.expect(:responses, mock_responses_api)
-
-    api = ClaudeSwarm::OpenAIResponses.new(
-      openai_client: mock_openai_client,
-      mcp_client: mock_mcp_client,
-      available_tools: [],
-      logger: mock_logger,
-      instance_name: "test",
-      model: "o1-pro",
-      reasoning_effort: "high",
-    )
-
-    mock_response = {
-      "id" => "resp_123",
-      "output" => [{
-        "type" => "message",
-        "content" => [{
-          "type" => "text",
-          "text" => "O1-pro response with reasoning",
+      # Mock the create call - verify reasoning is NOT included for GPT models
+      mock_response = {
+        "id" => "resp_123",
+        "output" => [{
+          "type" => "message",
+          "content" => [{
+            "type" => "text",
+            "text" => "Hello from GPT-4o responses API!",
+          }],
         }],
-      }],
-    }
+      }
 
-    mock_responses_api.expect(:create, mock_response) do |params|
-      params[:parameters][:reasoning] == { effort: "high" } &&
-        params[:parameters][:model] == "o1-pro" &&
-        !params[:parameters].key?(:temperature) # Also verify no temperature
+      mock_responses_api.expect(:create, mock_response) do |params|
+        !params[:parameters].key?(:reasoning) &&
+          params[:parameters][:model] == "gpt-4o"
+      end
+
+      result = api.execute("Hello")
+
+      assert_equal("Hello from GPT-4o responses API!", result)
+
+      mock_openai_client.verify
+      mock_responses_api.verify
     end
 
-    result = api.execute("Test")
+    def test_chat_completion_includes_reasoning_effort_only_for_o_series
+      # Test that reasoning_effort is included for O series model
+      mock_logger = Minitest::Mock.new
+      def mock_logger.info(*args); end
+      def mock_logger.error(*args); end
+      def mock_logger.log(*args); end
 
-    assert_equal("O1-pro response with reasoning", result)
+      mock_mcp_client = Minitest::Mock.new
+      mock_openai_client = Minitest::Mock.new
 
-    mock_openai_client.verify
-    mock_responses_api.verify
+      api = ClaudeSwarm::OpenAI::ChatCompletion.new(
+        openai_client: mock_openai_client,
+        mcp_client: mock_mcp_client,
+        available_tools: [],
+        logger: mock_logger,
+        instance_name: "test",
+        model: "o3",
+        reasoning_effort: "low",
+      )
+
+      mock_response = {
+        "choices" => [{
+          "message" => {
+            "role" => "assistant",
+            "content" => "O3 response with reasoning",
+          },
+        }],
+      }
+
+      mock_openai_client.expect(:chat, mock_response) do |params|
+        params[:parameters][:reasoning_effort] == "low" &&
+          params[:parameters][:model] == "o3" &&
+          !params[:parameters].key?(:temperature) # Also verify no temperature
+      end
+
+      result = api.execute("Test")
+
+      assert_equal("O3 response with reasoning", result)
+
+      mock_openai_client.verify
+    end
+
+    def test_responses_api_includes_reasoning_only_for_o_series
+      # Test that reasoning is included for O series model
+      mock_logger = Minitest::Mock.new
+      def mock_logger.info(*args); end
+      def mock_logger.error(*args); end
+      def mock_logger.log(*args); end
+
+      mock_mcp_client = Minitest::Mock.new
+      mock_openai_client = Minitest::Mock.new
+      mock_responses_api = Minitest::Mock.new
+
+      mock_openai_client.expect(:responses, mock_responses_api)
+
+      api = ClaudeSwarm::OpenAI::Responses.new(
+        openai_client: mock_openai_client,
+        mcp_client: mock_mcp_client,
+        available_tools: [],
+        logger: mock_logger,
+        instance_name: "test",
+        model: "o1-pro",
+        reasoning_effort: "high",
+      )
+
+      mock_response = {
+        "id" => "resp_123",
+        "output" => [{
+          "type" => "message",
+          "content" => [{
+            "type" => "text",
+            "text" => "O1-pro response with reasoning",
+          }],
+        }],
+      }
+
+      mock_responses_api.expect(:create, mock_response) do |params|
+        params[:parameters][:reasoning] == { effort: "high" } &&
+          params[:parameters][:model] == "o1-pro" &&
+          !params[:parameters].key?(:temperature) # Also verify no temperature
+      end
+
+      result = api.execute("Test")
+
+      assert_equal("O1-pro response with reasoning", result)
+
+      mock_openai_client.verify
+      mock_responses_api.verify
+    end
   end
 end
